* make undefined procdure handling safe for stack-shifter

+++ src/pl-prims.c
@@ -1269,7 +1269,7 @@ start:
     }
 
 #ifdef O_SHIFT_STACKS
-    if ( roomStack(global) < 2 * sizeof(word) )
+    if ( roomStack(global) < (intptr_t)(2 * sizeof(word)) )
     { if ( !growStacks(environment_frame, NULL, NULL, 0, 2 * sizeof(word), 0) )
 	return -1;
       p = valTermRef(t);
